requirements.txt
cryptography==3.4.8
google_api_python_client==2.58.0
google_auth_oauthlib==0.5.2
protobuf==4.21.5
psutil==5.9.1
pyperclip==1.8.2
PyQt5==5.15.7
